+++ packages/stylelint/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## stylelint@0.0.15 (2026-04-01)
+
+- Use `@tofrankie/postcss-wxml` to parse WXML files
+- Disable `no-empty-source` rule in `miniprogram` preset
+
 ## stylelint@0.0.14 (2026-03-30)
 
 - Disable `no-empty-source` rule in `miniprogram` preset